requirements.txt
django>=4.0.0,<4.1.0
algoliasearch-django>=3.0,<4.0
djangorestframework
djangorestframework-simplejwt
pyyaml
requests
django-cors-headers